backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
pydantic[email]
email-validator

